Extract exception classes
