+++ lib/elk.rb
@@ -13,17 +13,6 @@ module Elk
   # API version supported
   API_VERSION = "a1"
 
-  # When the authentication can't be done
-  class AuthError < RuntimeError; end
-  # Raised when the API server isn't working
-  class ServerError < RuntimeError; end
-  # Generic exception when 46elks API gives a response Elk can't parse
-  class BadResponse < RuntimeError; end
-  # Generic exception when Elk calls 46elk API the wrong way
-  class BadRequest < RuntimeError; end
-  # Raised when required paremeters are omitted
-  class MissingParameter < RuntimeError; end
-
   class << self
 
     extend Forwardable
@@ -59,6 +48,7 @@ end
 
 # Internal
 require_relative "elk/util"
+require_relative "elk/error"
 require_relative "elk/version"
 require_relative "elk/client"
 require_relative "elk/number"
